refactor(tool): move documents cache to helpers
- write tools cannot reach the read-side docs cache to drop stale entries, so cross-module invalidation is impossible.
- promote the cache into tools/_helpers with public get/store/invalidate helpers so both modules share one entry point.
