+++ src/mcp_server_polarion/tools/_helpers.py
@@ -12,6 +12,8 @@ inside ``read_document_parts``) and lives in ``tools.read``.
 
 from __future__ import annotations
 
+import time
+from dataclasses import dataclass
 from typing import Final, Literal, TypedDict, cast
 from urllib.parse import quote
 
@@ -633,6 +635,49 @@ def build_document_comment(item: dict[str, object]) -> DocumentComment:
     )
 
 
+# Document-discovery TTL cache (in-process, keyed by project_id).
+# A short TTL keeps paginated `list_documents` calls cheap while ensuring
+# newly-created documents appear within ~1 minute. ``create_document`` calls
+# ``invalidate_documents_cache`` to clear the entry immediately on write.
+_CACHE_TTL_SECONDS: Final[float] = 60.0
+
+
+@dataclass(frozen=True, slots=True)
+class _DocCacheEntry:
+    expires_at: float
+    documents: tuple[tuple[str, str], ...]
+
+
+_documents_cache: dict[str, _DocCacheEntry] = {}
+
+
+def get_cached_documents(project_id: str) -> list[tuple[str, str]] | None:
+    """Return the cached document list for *project_id* or ``None``."""
+    entry = _documents_cache.get(project_id)
+    if entry is None:
+        return None
+    if time.monotonic() >= entry.expires_at:
+        _documents_cache.pop(project_id, None)
+        return None
+    return list(entry.documents)
+
+
+def store_cached_documents(
+    project_id: str,
+    documents: list[tuple[str, str]],
+) -> None:
+    """Cache *documents* for *project_id* for ``_CACHE_TTL_SECONDS``."""
+    _documents_cache[project_id] = _DocCacheEntry(
+        expires_at=time.monotonic() + _CACHE_TTL_SECONDS,
+        documents=tuple(documents),
+    )
+
+
+def invalidate_documents_cache(project_id: str) -> None:
+    """Drop the cached document list for *project_id*, if any."""
+    _documents_cache.pop(project_id, None)
+
+
 __all__: list[str] = [
     "DEFAULT_PAGE_SIZE",
     "DOCUMENT_COMMENT_LIST_FIELDS",
@@ -652,13 +697,16 @@ __all__: list[str] = [
     "extract_relationship_ids",
     "extract_short_id",
     "extract_total_count",
+    "get_cached_documents",
     "get_client",
     "has_links_next",
+    "invalidate_documents_cache",
     "merge_custom_fields",
     "parse_hyperlinks",
     "parse_work_item_detail",
     "parse_work_item_summaries",
     "safe_str",
     "split_module_id",
+    "store_cached_documents",
     "summary_to_back_link",
 ]
