+++ src/mcp_server_polarion/tools/read.py
@@ -18,7 +18,6 @@ Body fields use two different formats depending on the tool's purpose:
 from __future__ import annotations
 
 import re
-import time
 from dataclasses import dataclass
 from typing import Final, Literal, cast
 
@@ -62,11 +61,13 @@ from mcp_server_polarion.tools._helpers import (
     extract_relationship_id,
     extract_short_id,
     extract_total_count,
+    get_cached_documents,
     get_client,
     parse_work_item_detail,
     parse_work_item_summaries,
     safe_str,
     split_module_id,
+    store_cached_documents,
     summary_to_back_link,
 )
 from mcp_server_polarion.utils import html_to_markdown
@@ -466,42 +467,6 @@ def _get_module_id(item: object) -> str:
     return extract_relationship_id(relationships, "module")
 
 
-# Document-discovery TTL cache (in-process, keyed by project_id).
-# A short TTL keeps paginated `list_documents` calls cheap while ensuring
-# newly-created documents appear within ~1 minute. Future write tools can
-# invalidate by popping the project_id key.
-_CACHE_TTL_SECONDS: Final[float] = 60.0
-
-
-@dataclass(frozen=True, slots=True)
-class _DocCacheEntry:
-    expires_at: float
-    documents: tuple[tuple[str, str], ...]
-
-
-_documents_cache: dict[str, _DocCacheEntry] = {}
-
-
-def _get_cached_documents(project_id: str) -> list[tuple[str, str]] | None:
-    entry = _documents_cache.get(project_id)
-    if entry is None:
-        return None
-    if time.monotonic() >= entry.expires_at:
-        _documents_cache.pop(project_id, None)
-        return None
-    return list(entry.documents)
-
-
-def _store_cached_documents(
-    project_id: str,
-    documents: list[tuple[str, str]],
-) -> None:
-    _documents_cache[project_id] = _DocCacheEntry(
-        expires_at=time.monotonic() + _CACHE_TTL_SECONDS,
-        documents=tuple(documents),
-    )
-
-
 async def _discover_documents(
     client: PolarionClient,
     project_id: str,
@@ -509,8 +474,8 @@ async def _discover_documents(
     """Discover all unique (space_id, document_name) pairs via linear scan.
 
     Iterates every heading-workitem page (page_size=100) and accumulates
-    unique ``module`` relationship IDs. Results are cached for
-    ``_CACHE_TTL_SECONDS`` so paginated callers reuse the discovery.
+    unique ``module`` relationship IDs. Results are TTL-cached in
+    ``tools._helpers`` so paginated callers reuse the discovery.
 
     Args:
         client: Active ``PolarionClient`` instance.
@@ -519,7 +484,7 @@ async def _discover_documents(
     Returns:
         Sorted list of (space_id, document_name) tuples.
     """
-    cached = _get_cached_documents(project_id)
+    cached = get_cached_documents(project_id)
     if cached is not None:
         return cached
 
@@ -552,7 +517,7 @@ async def _discover_documents(
         page_number += 1
 
     sorted_docs = sorted(documents)
-    _store_cached_documents(project_id, sorted_docs)
+    store_cached_documents(project_id, sorted_docs)
     return sorted_docs
